source/electrical_heating/pyproject.toml
[project]
name = "opengeh_electrical_heating"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-core==1.32.0",
    "azure-monitor-opentelemetry==1.6.4",
    "configargparse==1.7.0",
    "delta-spark==3.2.0",
    "opengeh-pyspark @ git+https://github.com/Energinet-DataHub/opengeh-python-packages@3.1.1#subdirectory=source/pyspark_functions",
    "opengeh-telemetry @ git+https://github.com/Energinet-DataHub/opengeh-python-packages@cbang/simplify_entry_point2#subdirectory=source/telemetry",
    "pydantic-settings>=2.7.1",
    "pyspark==3.5.1",
    "python-dateutil==2.8.2",
    "tomli>=2.2.1",
]

[project.scripts]
execute = "opengeh_electrical_heating.entry_point:execute"

[dependency-groups]
dev = [
    "commitizen>=4.1.0",
    "mypy>=1.14.1",
    "opengeh-testcommon @ git+https://github.com/Energinet-DataHub/opengeh-python-packages@3.1.1#subdirectory=source/testcommon",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.9.1",
    "databricks-sdk==0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
# -s ensures that print("<text>") is written to output.
addopts = "--cov-report=term --cov-report=xml:coverage.xml --cov=opengeh_electrical_heating -ra -vv -s"

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.ruff]
line-length = 120
show-fixes = true

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["D", "D401", "E", "F", "I", "T201", "UP"]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D106",
    "D107",
    "D417",
    "E501",
    "UP006",
    "UP007",
    "UP035",
    "F401",
]
exclude = ["notebooks/*"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.uv.sources]
opengeh-telemetry = { git = "https://github.com/Energinet-DataHub/opengeh-python-packages", subdirectory = "source/telemetry", branch = "cbang/simplify_entry_point2" }
opengeh-testcommon = { git = "https://github.com/Energinet-DataHub/opengeh-python-packages", subdirectory = "source/testcommon", rev = "3.1.1" }
opengeh-pyspark = { git = "https://github.com/Energinet-DataHub/opengeh-python-packages", subdirectory = "source/pyspark_functions", rev = "3.1.1" }
